Typescript migration - fixing errors - seeking team help
