+++ package.json
@@ -4,9 +4,11 @@
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js",
-    "dev_start": "nodemon server.js",
-    "lint": "eslint ."
+    "start": "node dist/server.js",
+    "dev_start": "nodemon dist/server.js",
+    "dev": "ts-node src/server.ts",
+    "build": "tsc",
+    "lint": "eslint --ext .ts src"
   },
   "repository": {
     "type": "git",
@@ -37,9 +39,16 @@
     "@eslint/js": "^9.19.0",
     "@types/bcryptjs": "^2.4.6",
     "@types/express": "^5.0.0",
+    "@types/express-session": "^1.18.1",
     "@types/jsonwebtoken": "^9.0.9",
     "@types/mongoose": "^5.11.96",
     "@types/node": "^22.13.10",
+    "@types/passport": "^1.0.17",
+    "@types/passport-google-oauth20": "^2.0.16",
+    "@types/swagger-jsdoc": "^6.0.4",
+    "@types/swagger-ui-express": "^4.1.8",
+    "@typescript-eslint/eslint-plugin": "^8.26.0",
+    "@typescript-eslint/parser": "^8.26.0",
     "eslint": "^9.19.0",
     "eslint-config-prettier": "^10.0.1",
     "eslint-plugin-prettier": "^5.2.3",
